"changes made"

+++ buylist/models.py
@@ -40,7 +40,7 @@ class StarcityBuylist(models.Model):
 
 class StoreCredit(models.Model):
     name = models.CharField(max_length=255, default='')
-    total = models.IntegerField(default=0)
+    total = models.DecimalField(max_digits=12, decimal_places=2, default=0)
     entries = models.IntegerField(default=0)
 
     def __str__(self):
